editied the hubconf file
